+++ src/background/background.ts
@@ -27,7 +27,7 @@ async function toDoOnFirstInstall(details: InstalledDetails) {
         if (details.reason === 'install') {
         // Open the welcome page
         api.tabs.create({
-            url: api.runtime.getURL('dist/popup/welcome.html')
+            url: api.runtime.getURL('dist/popup/settings.html')
         });
 
         // Find all open YouTube and YouTube No-Cookie tabs and refresh them